Microservices/SentimentAPI/requirements.txt
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
click==8.2.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.3
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.1
kaggle==1.7.4.2
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
protobuf==6.30.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.7
text-unidecode==1.3
tzdata==2025.2
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
yt-dlp==2025.2.19
